* Attempt to fix Trait exp overflow/softcap All six traits' get methods cast the trait exp to short; this will cause the value to overflow into -32678 in the progress towards trait 55 and effectively made trait levels 55+ unobtainable. To prevent the *very* unlikely case of int overflow into -2.1b from ocurring, a hard cap has been introduced to addTraitExp to stop that character from receiving anymore trait exp for a specific maxed trait. Consequently, they will not see any trait exp messages for that trait henceforth. In my testing, the trait exp might not show changes for the first instance of trait exp gain. Try ccing/relogging to get the client to recalculate traits. * Implement Appearance Reset Coupon and Trait Boost Potion * Script The Light of Hope (Throne of Masters quest) The quest still can't be accepted normally; a glance at Check.img.xml implies that being able to accept it is beholden to a date range like event quests. I'm not sure how to get around this beyond force starting the quest through another script or outright giving the chair through other means.
(cherry picked from commit 1b6d84b8dc77d99f06d1cd19ab5884572c9f0af4)

+++ src/main/java/net/swordie/ms/client/character/CharacterStat.java
@@ -140,8 +140,8 @@ public class CharacterStat {
         return (short) level;
     }
 
-    public short getCharismaExp() {
-        return (short) charismaExp;
+    public int getCharismaExp() {
+        return charismaExp;
     }
 
     public short getLuk() {
@@ -228,12 +228,12 @@ public class CharacterStat {
         return worldIdForLog;
     }
 
-    public short getCharmExp() {
-        return (short) charmExp;
+    public int getCharmExp() {
+        return charmExp;
     }
 
-    public short getCraftExp() {
-        return (short) craftExp;
+    public int getCraftExp() {
+        return craftExp;
     }
 
     public int getAlbaActivityID() {
@@ -252,8 +252,8 @@ public class CharacterStat {
         return albaDuration;
     }
 
-    public short getInsightExp() {
-        return (short) insightExp;
+    public int getInsightExp() {
+        return insightExp;
     }
 
     public int getAlbaSpecialReward() {
@@ -280,12 +280,12 @@ public class CharacterStat {
         return pvpPoint;
     }
 
-    public short getSenseExp() {
-        return (short) senseExp;
+    public int getSenseExp() {
+        return senseExp;
     }
 
-    public short getWillExp() {
-        return (short) willExp;
+    public int getWillExp() {
+        return willExp;
     }
 
     public long getPosMap() {
